Fixed rendering

+++ engine/core/include/graphics/buffers/vertex_buffer.h
@@ -61,7 +61,7 @@ public:
         return stride;
     }
 
-    uint32 size() {
+    size_t size() {
         return elements.size();
     }
 };
